+++ src/components/Product/Detail.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { useParams } from 'react-router-dom'
+
+export default function ProductDetail() {
+  const { id } = useParams();
+  
+  return (
+    <div>Product detail {id}</div>
+  )
+};
