--- src/components/Product/ProductDetail.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function ProductDetail() {
-  return (
-    <div>ProductDetail</div>
-  )
-}
