--- src/components/Product/ProductEdit.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function ProductEdit() {
-  return (
-    <div>ProductEdit</div>
-  )
-}
