--- src/components/Product/ProductList.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import Herosection from '../Herosection'
-
-export default function ProductList() {
-  return (
-    <React.Fragment>
-      <Herosection />
-      <div>ProductList</div>
-    </React.Fragment>
-  )
-}
